+++ src/types/data.ts
@@ -4,6 +4,12 @@ export interface BaseProperties {
   unused?: boolean;
 }
 
+export interface EquipmentStats {
+  attack: number;
+  defence: number;
+  magic: number;
+}
+
 export interface WithOverrides<T, A = unknown> {
   getOverrides?: (args: A) => Partial<T>;
 }